requirements-minimal.txt
blinker==1.7.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Flask==3.0.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.0.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
openai==0.8.0
pydantic==1.9.0
PyJWT==2.7.0
PyMySQL==1.1.1
pycparser==2.22
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.48
tqdm==4.64.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.14
Werkzeug==3.0.1
